Update initDataDB
	la classe acquiert une nouvelle méthode InitApp
	qui prend l'annotation de scheduled task
	InitDb n'est désormais appalée que sous certaines condition

	reprise complete du jeu de donnée par défaut
	ajout de méthodes procédurales pour la création de données
	Mais le jeux de donnée est devenu tellement grand > 10000 entrées
	que l'application met > 2 min à démarrer
Après modification l'application mais 10s si la base de données est déjà
en place et que l'entrée initdb contient la valeurb true

Création ApplicationParams DAO
			 ApplicationParamsService
			 ApplicationParamsServiceImpl
			 ApplicationParamsRepository
			 le role de cette classe est de gérer
			 des paramètres de l'application en base de données

Refactoring  de classe DAO et tous les test associé
	BigDecimals remplacés par des float
	la précision trop grande des big decimals les rend imprévisible pour
	des valeurs monétaires
	Les test unitaires montre une précision satisfaisante
	voir dans ScheduledTaskTest
	ou dans NatureTest

Correctif dans MissionControler
	il emploie désormais les ErrorsMessages
	et renvoie des badCrédential exception

+++ src/main/java/diginamic/gdm/dto/MissionDTO.java
@@ -33,7 +33,7 @@ public class MissionDTO implements DTO<Mission> {
     /** arrivalCity */
     private CityDTO arrivalCity;
     /** bonus */
-    private BigDecimal bonus;
+    private float bonus;
     /** transport */
     private Transport transport;
     /** status */
